ByteArray: move to o.b.secp.internal package
Remove ByteArray from our public API.
