+++ secp-api/src/main/java/org/bitcoinj/secp/internal/SecpXOnlyPubKeyImpl.java
@@ -15,7 +15,6 @@
  */
 package org.bitcoinj.secp.internal;
 
-import org.bitcoinj.secp.ByteArray;
 import org.bitcoinj.secp.SecpFieldElement;
 import org.bitcoinj.secp.SecpXOnlyPubKey;
